Registration+ValidToken

+++ API/Controllers/TShirtController.cs
@@ -3,8 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using API.DTO;
+using API.HelperToQuery;
 using API.Interfaces;
 using API.Mappers;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace API.Controllers
@@ -20,9 +22,15 @@ namespace API.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll ()
+        [Authorize]
+        public async Task<IActionResult> GetAll ([FromQuery] QueryObject queryObject)
         {
-            var tShirt = await _tShirtRepo.GetAllAsync();
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var tShirt = await _tShirtRepo.GetAllAsync(queryObject);
 
             var tshirtDto = tShirt.Select(t => t.IntoTShirtDto());
 
@@ -32,6 +40,11 @@ namespace API.Controllers
         [HttpGet("{id:int}")]
         public async Task <IActionResult> GetById ([FromRoute]int id)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var tShirt = await _tShirtRepo.GetByIdAsync(id);
 
             if(tShirt == null)
@@ -43,8 +56,14 @@ namespace API.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public async Task <IActionResult> Create ([FromBody] CreateTShirtDto tshirtDTo)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var tShirtModel = tshirtDTo.ToUpdateDto();
             await _tShirtRepo.CreateAsync(tShirtModel);
             return CreatedAtAction(nameof(GetById), new {id = tShirtModel.Id}, tShirtModel.IntoTShirtDto());
@@ -54,6 +73,11 @@ namespace API.Controllers
         [Route("{id:int}")]
         public async Task<IActionResult> Update ([FromRoute] int id, [FromBody] UpdateTshirtDto updateTshirtDto)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var tShirtModel = await _tShirtRepo.UpdateAsync(id, updateTshirtDto);
             if(tShirtModel == null)
             {
@@ -64,13 +88,18 @@ namespace API.Controllers
 
         [HttpDelete]
         [Route("{id:int}")]
-        public async Task<IActionResult> Delete ([FromRoute] int id)
+        public async Task<IActionResult> Delete([FromRoute] int id)
         {
-            var TshirtExists = await _tShirtRepo.DeleteAsync(id);
-
-            if(TshirtExists == null)
+            if (!ModelState.IsValid)
             {
-                return NotFound("i din't find a corrresponding ID sir");
+                return BadRequest(ModelState);
+            }
+
+            var deletedTShirt = await _tShirtRepo.DeleteAsync(id);
+
+            if (deletedTShirt == null)
+            {
+                return NotFound("No corresponding TShirt found with the specified ID.");
             }
 
             return NoContent();
